Create main architecture for user profile fragment
