--- app/src/main/java/pro/fateeva/chuchasgithub/UserProfileFragment.kt
@@ -1,30 +1,0 @@
-package pro.fateeva.chuchasgithub
-
-import androidx.lifecycle.ViewModelProvider
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-class UserProfileFragment : Fragment() {
-
-    private lateinit var viewModel: UserProfileViewModel
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.user_profile_fragment, container, false)
-    }
-
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(UserProfileViewModel::class.java)
-        // TODO: Use the ViewModel
-    }
-
-    companion object {
-        fun newInstance() = UserProfileFragment()
-    }
-}
