--- app/src/main/java/pro/fateeva/chuchasgithub/UserProfileViewModel.kt
@@ -1,7 +1,0 @@
-package pro.fateeva.chuchasgithub
-
-import androidx.lifecycle.ViewModel
-
-class UserProfileViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
