+++ app/src/main/java/pro/fateeva/chuchasgithub/data/UserRepository.kt
@@ -3,7 +3,7 @@ package pro.fateeva.chuchasgithub.data
 import pro.fateeva.chuchasgithub.domain.entities.User
 
 interface UserRepository {
-    fun getUsersList() : List<User>
+    fun getUserList() : List<User>
     fun addUser(user: User)
     fun getUser(id: Int): User?
     fun deleteUser(user: User)
