+++ app/src/main/java/pro/fateeva/chuchasgithub/domain/UserProfileUseCase.kt
@@ -1,0 +1,4 @@
+package pro.fateeva.chuchasgithub.domain
+
+interface UserProfileUseCase {
+}
