+++ app/src/main/java/pro/fateeva/chuchasgithub/domain/UserProfileUseCaseImpl.kt
@@ -1,0 +1,6 @@
+package pro.fateeva.chuchasgithub.domain
+
+import pro.fateeva.chuchasgithub.data.UserRepository
+
+class UserProfileUseCaseImpl(private val repository: UserRepository) : UserProfileUseCase {
+}
