+++ app/src/main/java/pro/fateeva/chuchasgithub/ui/userProfile/UserProfileViewModel.kt
@@ -1,0 +1,8 @@
+package pro.fateeva.chuchasgithub.ui.userProfile
+
+import androidx.lifecycle.ViewModel
+import pro.fateeva.chuchasgithub.domain.UserProfileUseCase
+
+class UserProfileViewModel : ViewModel() {
+    lateinit var userProfileUseCase: UserProfileUseCase
+}
