+++ app/src/main/res/layout/user_profile_fragment.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".UserProfileFragment">
+    tools:context=".ui.userProfile.UserProfileFragment">
 
     <ImageView
         android:id="@+id/photoImageView"
